ad14n_ad15n_ac104n-release_v1.6.0 SDK等级： B 发布时间：2022年12月12日 发布版本：ad14n-ad15n_ac104n-release_V1.6.0 发布说明： 1、新增ac104n音箱sdk； 以下改动对应”ad140-release_v1.4.0”和”ad150-release_v1.4.0” 2、修复外挂flash操作大容量flash（大于16MB）失败问题； 3、修复vfs_mount失败时不关闭文件系统的bug； 4、修复msg带参数消息接受失败的bug； 5、修复退出midi解码模式没有清除解码回调导致midi琴无声问题； 6、优化A格式解码断点大小，节省空间； 7、优化midi解码默认采样率，从48K改为32K； 8、新增内置flash代码资源区硬件写保护； 9、新增内置flash软件驱动擦写限制，不操作代码资源区； 10、新增powerdown vddio电压保持功能； 11、新增syd文件系统计算文件CRC16的接口（可打开dir文件，仅用于校验）； 12、更新midi库，修复midi琴note_off在音色无包络时没有衰减的问题； 13、更新midi库，新增midi琴note_off衰减参数，可选择note_off是否回调音符结束； 14、更新midi库，新增midi琴查询指定通道正在播放的key； 15、更新Notch_Howling陷波啸叫抑制算法库，优化运算速度；
文档更新：
1、新增文档《AD14N_AD15N_AC104N内置FLASH添加写保护说明.pdf》；
2、更新文档《杰理AD14N_AD15N_MIDI应用说明文档.pdf》；
3、更新文档《杰理AD14N_AD15N_AC104N SDK手册.pdf》；
4、更新了芯片规格书《AD146A_Datasheet V1.1.pdf》；
5、更新了硬件原理图；

工具版本：
1、一拖二烧写器：V2.26.8，通过软件上远程更新；
2、一拖八烧写器：V3.1.14，通过软件上远程更新；
3、音频文件转换工具：V1.2.17，在gitee上下载；
4、JLmidi_new_0706工具：该工具只能用于此发布版本以及之后的SDK，在gitee上下载，gitee上会保留旧工具（JLmidi_60）；

+++ sdk/include_lib/common/tab_read.h
@@ -10,7 +10,7 @@ typedef struct _rtab_obj {
     u32   cnt;
 } rtab_obj;
 
-void tab_init(rtab_obj *stab, void *tab, u16 size);
+void tab_init(rtab_obj *stab, void *tab, u32 size);
 u32 tab_read(void *buff, rtab_obj *stab, u32 len);
